Create Grid module under board
This creates a grid module under the board module. The grid module
contains the grid struct which is used to encapsulate all board grid
interactions to seperate concerns from the board which is more about
orchestration. It also contains a grid::new() function to create a
Grid. This also moved the player module under the board module.
